2025-2026/PMPm-22/master-thesis/Pastushuk_Olha/backend/requirements.txt
fastapi
uvicorn
torch --index-url https://download.pytorch.org/whl/cpu
librosa
numpy<2.0.0
webrtcvad-wheels
python-multipart